feat(admin,core): image sub-fields in repeaters with media picker
Repeater rows rendered non-scalar sub-fields as plain text inputs.
Adds case 'image' to the sub-field renderer (reusing ImageFieldRenderer,
extracted from ContentEditor to avoid a circular import), the builder
type option, and extends REPEATER_SUB_FIELD_TYPES + the API Zod enum
(which was also missing the documented 'url' type).

Closes #1424
